requirements.txt
fastapi[standard]>=0.115.12
mcp>=1.6.0
python-dotenv>=1.1.0
requests>=2.32.3